+++ R/cyto_gateTemplate.R
@@ -113,7 +113,10 @@ cyto_gateTemplate.GatingHierarchy <- function(x) {
             )[[1]],
             parent = gt$parent[ind[z]],
             gate = gates[[z]],
-            channels = gt$dims[ind[z]]
+            channels = strsplit(
+              gt$dims[ind[z]],
+              ","
+            )[[1]]
           )
         }
       ),